test/fixtures/wpt/fetch/nosniff/importscripts.html

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
 async_test(function(t) {
   var w = new Worker("importscripts.js")
   w.onmessage = t.step_func(function(e) {
     if(e.data == "END")
       t.done()
     else
       assert_equals(e.data, "PASS")
   })
 }, "Test importScripts()")
</script>
